Added better error checking for API requests. Error checks now use the http response code.

+++ controller/tdaApi.php
@@ -32,67 +32,76 @@ class TdaApi
     }
 
     /** Create and save brand new Refresh and Access tokens using a Permission Code. */
-    public function createTokens(string $accountId, string $permissionCode): void
+    public function createTokens(string $accountId, string $permissionCode): bool
     {
         $accountInfo = $this->mySql->read('tda_api', ['account_id' => $accountId])[0];
         $tokenRequest = $this->tdaApiRequest->newTokens($permissionCode, $accountInfo->consumerKey, $accountInfo->redirectUri);
-        if (isset($tokenRequest->error)) {
-            $this->log->save('tokens', 'Account#: ' . $accountId . ' - Error creating new tokens: ' . $tokenRequest->error);
+        if ($tokenRequest->httpdCode !== 200) {
+            $this->log->save('tokens', 'Account#: ' . $accountId . ' - Error creating new tokens: ' . $tokenRequest->response->response->error);
+            return FALSE;
         } else {
-            $this->saveTokens($accountId, $tokenRequest->access_token, $tokenRequest->expires_in, $tokenRequest->refresh_token, $tokenRequest->refresh_token_expires_in);
+            $this->saveTokens($accountId, $tokenRequest->response->access_token, $tokenRequest->response->expires_in, $tokenRequest->response->refresh_token, $tokenRequest->response->refresh_token_expires_in);
             $this->log->save('tokens', 'Account#: ' . $accountId . ' - Successfully created new tokens.');
+            return TRUE;
         }
     }
 
-    /** Update and save expired tokens. */
-    public function updateTokens(string $accountId): void
+    /** Update and save expired tokens. Return FALSE if there was an issue updating tokens.*/
+    public function updateTokens(string $accountId): bool
     {
         $accountInfo = $this->mySql->read('tda_api', ['account_id' => $accountId])[0];
         if (time() < $accountInfo->accessTokenExpiration) {
             $this->log->save('tokens', 'Account#: ' . $accountId . ' - Tokens are not expired.');
+            return TRUE;
         } else {
             if (time() >= $accountInfo->refreshTokenExpiration) {
                 // Refresh token is expired. Update both Refresh and Access tokens.
                 $tokenRequest = $this->tdaApiRequest->refreshToken($accountInfo->refreshToken, $accountInfo->consumerKey);
-                if (isset($tokenRequest->error)) {
-                    $this->log->save('tokens', 'Account#: ' . $accountId . ' - Error updating refresh token: ' . $tokenRequest->error);
+                if ($tokenRequest->httpdCode !== 200) {
+                    $this->log->save('tokens', 'Account#: ' . $accountId . ' - Error updating refresh token: ' . $tokenRequest->response->error);
+                    return FALSE;
                 } else {
-                    $this->saveTokens($accountId, $tokenRequest->access_token, $tokenRequest->expires_in, $tokenRequest->refresh_token, $tokenRequest->refresh_token_expires_in);
+                    $this->saveTokens($accountId, $tokenRequest->response->access_token, $tokenRequest->response->expires_in, $tokenRequest->response->refresh_token, $tokenRequest->response->refresh_token_expires_in);
                     $this->log->save('tokens', 'Account#: ' . $accountId . ' - Successfully updated refresh and access tokens.');
+                    return TRUE;
                 }
             } else {
                 // Only the access token is expired. Update just the Access token.
                 $tokenRequest = $this->tdaApiRequest->accessToken($accountInfo->refreshToken, $accountInfo->consumerKey);
-                if (isset($tokenRequest->error)) {
-                    $this->log->save('tokens', 'Account#: ' . $accountId . ' - Error updating access token: ' . $tokenRequest->error);
+                if ($tokenRequest->httpdCode !== 200) {
+                    $this->log->save('tokens', 'Account#: ' . $accountId . ' - Error updating access token: ' . $tokenRequest->response->error);
+                    return FALSE;
                 } else {
-                    $this->saveTokens($accountId, $tokenRequest->access_token, $tokenRequest->expires_in);
+                    $this->saveTokens($accountId, $tokenRequest->response->access_token, $tokenRequest->response->expires_in);
                     $this->log->save('tokens', 'Account#: ' . $accountId . ' - Successfully updated access token.');
+                    return TRUE;
                 }
             }
         }
     }
 
     /** Download and save transactions for a specific date range. */
-    public function updateTransactions(string $accountId, ?string $startDate = NULL, ?string $endDate = NULL): void
+    public function updateTransactions(string $accountId, ?string $startDate = NULL, ?string $endDate = NULL): bool
     {
-        $accountInfo = $this->mySql->read('tda_api', ['account_id' => $accountId])[0];
-
-        // Clear old pending transactions, and download transactions for the specified date range.
-        $this->mySql->truncate('transactions_pending');
         $today = date("Y-m-d");
         $startDate = $startDate ?? $today;
         $endDate = $endDate ?? $today;
-        $transactions = $this->tdaApiRequest->transactions($accountInfo->accessToken, $accountInfo->accountNumber, $startDate, $endDate);
-        $transactions = array_reverse($transactions);
+        $accountInfo = $this->mySql->read('tda_api', ['account_id' => $accountId])[0];
+        $transactionsRequest = $this->tdaApiRequest->transactions($accountInfo->accessToken, $accountInfo->accountNumber, $startDate, $endDate);
+        if ($transactionsRequest->httpdCode !== 200) {
+            $this->log->save('transactions', 'Error downloading transactions for account' . $accountId . ': ' . $transactionsRequest->response->error);
+            return FALSE;
+        }
 
         // Flatten each transacation and store in the database. Record transaction type (duplicate, pending, completed), and transaction id/order id.
+        $transactions = array_reverse($transactionsRequest->response);
+        $this->mySql->truncate('transactions_pending');
         $transactionsUpdated = [];
         $transactionsDuplicates = [];
         $transactionsPending = [];
         $transactionsErrors = [];
         foreach ($transactions as $transaction){
-            $transaction[] = ['account_id' => $accountId];
+            $transaction->account_id = $accountId;
             $processedTransaction = flatten($transaction);
             $transactionTable = (isset($transaction->transactionSubType)) ? 'transactions' : 'transactions_pending' ;
             $mySqlResponse = $this->mySql->create($transactionTable, $processedTransaction);
@@ -110,19 +119,33 @@ class TdaApi
                 }
             }
         }
-
-        // Save update information to the log.
         $logMessage = 'Account#: ' . $accountId . ' - Transactions Downloaded - New: ' . count($transactionsUpdated) . ' Duplicates: ' . count($transactionsDuplicates) . ' Pending: ' . count($transactionsPending) . ' Errors: ' . count($transactionsErrors);
         $this->log->save('transactions', $logMessage);
         if (count($transactionsErrors) > 0) {
             $this->log->save('transactions', 'Account#: ' . $accountId . ' - Failed to add transactions: ' . json_encode($transactionsErrors));
         }
+        return TRUE;
     }
 
     /** Download and save orders for a specified date range. */
-    public function updateOrders(): void
+    public function updateOrders(string $accountId, ?string $startDate = NULL, ?string $endDate = NULL)
     {
-        
+        $accountInfo = $this->mySql->read('tda_api', ['account_id' => $accountId])[0];
+        //echo '<pre>';
+        //var_dump($accountInfo);
+        //echo '</pre>';
+        $today = date("Y-m-d");
+        $startDate = $startDate ?? $today;
+        $endDate = $endDate ?? $today;
+        $orders = $this->tdaApiRequest->orders($startDate, $endDate, $accountInfo->accessToken);
+        return $orders;
+    }
+
+    public function getTdaAccount(string $accountId): object
+    {
+        $accountInfo = $this->mySql->read('tda_api', ['account_id' => $accountId])[0];
+        $tdaAccount = $this->tdaApiRequest->accountInfo($accountInfo->accessToken, $accountInfo->accountNumber);
+        return $tdaAccount;
     }
 
 }
